requirements.txt
arrow==1.3.0
asgiref==3.7.2
binaryornot==0.4.4
boto3==1.34.88
botocore==1.34.88
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cookiecutter==2.6.0
Django==5.0.3
django-admin-interface==0.28.6
django-appconf==1.0.6
django-colorfield==0.11.0
django-compressor==4.4
django-htmx==1.17.3
django-phonenumber-field==7.3.0
django-storages==1.14.2
django-suit @ https://github.com/darklow/django-suit/tarball/v2#sha256=b40c916de9528ca86bcccd42d1fd54e7773e5071d2bfd9efc576039c1a441c0b
django-unfold==0.20.5
django-widget-tweaks==1.5.0
DjangoSEO==1.0
idna==3.6
Jinja2==3.1.3
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
phonenumberslite==8.13.31
pillow==10.2.0
Pygments==2.17.2
python-dateutil==2.9.0.post0
python-decouple==3.6
python-slugify==8.0.4
PyYAML==6.0.1
rcssmin==1.1.1
requests==2.31.0
rich==13.7.1
rjsmin==1.2.1
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
stripe==5.5.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20240316
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
